requirements.txt
fastapi==0.115.9
uvicorn[standard]==0.38.0

# Keep Chroma modern; avoid backtracking
chromadb==1.2.1
tokenizers==0.20.3
# Let chromadb pull a compatible posthog (<6); do NOT pin 6.x
# posthog  (intentionally omitted)

openai==2.6.1
pdfminer.six==20250506
python-docx==1.2.0
python-pptx==1.0.2
openpyxl==3.1.5
PyYAML==6.0.3

pydantic[email]==2.12.3
pydantic-core==2.41.4
typing-inspection==0.4.2
annotated-types==0.7.0

bcrypt==4.2.1
passlib[bcrypt]==1.7.4
python-multipart
sqlalchemy==2.0.44

pydantic-settings>=2.0.0
psycopg2-binary
PyJWT==2.8.0